+++ lib/pages/home_page.dart
@@ -8,7 +8,6 @@ import 'package:localsend_app/init.dart';
 import 'package:localsend_app/pages/tabs/receive_tab.dart';
 import 'package:localsend_app/pages/tabs/send_tab.dart';
 import 'package:localsend_app/pages/tabs/settings_tab.dart';
-import 'package:localsend_app/pages/tabs/troubleshoot_tab.dart';
 import 'package:localsend_app/provider/selection/selected_sending_files_provider.dart';
 import 'package:localsend_app/theme.dart';
 import 'package:localsend_app/widget/responsive_builder.dart';
@@ -16,8 +15,7 @@ import 'package:localsend_app/widget/responsive_builder.dart';
 enum HomeTab {
   receive(Icons.wifi),
   send(Icons.send),
-  settings(Icons.settings),
-  troubleshoot(Icons.healing);
+  settings(Icons.settings);
 
   const HomeTab(this.icon);
 
@@ -31,8 +29,6 @@ enum HomeTab {
         return t.sendTab.title;
       case HomeTab.settings:
         return t.settingsTab.title;
-      case HomeTab.troubleshoot:
-        return t.troubleshootTab.title;
     }
   }
 }
@@ -142,7 +138,6 @@ class _HomePageState extends ConsumerState<HomePage> {
                             ReceiveTab(),
                             SendTab(),
                             SettingsTab(),
-                            TroubleshootTab(showTitle: true),
                           ],
                         ),
                         if (_dragAndDropIndicator)
@@ -170,13 +165,7 @@ class _HomePageState extends ConsumerState<HomePage> {
                 ? NavigationBar(
                     selectedIndex: sizingInformation.isMobile ? min(_currentTab.index, 2) : _currentTab.index,
                     onDestinationSelected: _goToPage,
-                    destinations: [
-                      HomeTab.receive,
-                      HomeTab.send,
-                      HomeTab.settings,
-                      if (sizingInformation.isTabletOrDesktop)
-                        HomeTab.troubleshoot,
-                    ].map((tab) {
+                    destinations: HomeTab.values.map((tab) {
                       return NavigationDestination(icon: Icon(tab.icon), label: tab.label);
                     }).toList(),
                   )